feat: separate 'push'
