+++ cli/flox/src/commands/mod.rs
@@ -7,6 +7,7 @@ mod general;
 mod init;
 mod install;
 mod list;
+mod push;
 mod search;
 mod uninstall;
 
@@ -444,7 +445,7 @@ impl LocalDevelopmentCommands {
 enum SharingCommands {
     /// Send an environment to FloxHub
     #[bpaf(command, footer("Run 'man flox-push' for more details."))]
-    Push(#[bpaf(external(environment::push))] environment::Push),
+    Push(#[bpaf(external(push::push))] push::Push),
     /// Pull an environment from FloxHub
     #[bpaf(command, footer("Run 'man flox-pull' for more details."))]
     Pull(#[bpaf(external(environment::pull))] environment::Pull),
